+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
-- enh(android): Add more telemetry (#435)
+- feat(sourcemaps): Automatically insert Sentry Webpack plugin (#432)
+- fix(android): Add support for unusual import statements (#440)
+- fix(wizard): Sort projects in project-selection step (#441)
+- enh(android): Add more telemetry (#435) 
 
 ## 3.11.0